onyo/requirements.txt
Django==1.9.6
djangorestframework==3.3.3
Markdown==2.6.6
requests==2.10.0
